add dry-fun flag for setup step

+++ packages/shipjs/src/step/setup/addScriptsToPackageJson.js
@@ -5,15 +5,17 @@ import { runPrettier } from '../../helper';
 import { info } from '../../color';
 import { print } from '../../util';
 
-export default ({ dir }) =>
-  runStep({ title: 'Adding scripts to package.json' }, () => {
-    const filePath = path.resolve(dir, 'package.json');
-    const json = JSON.parse(fs.readFileSync(filePath).toString());
-    json.scripts = json.scripts || {};
-    json.scripts['release:prepare'] = 'shipjs prepare';
-    json.scripts['release:trigger'] = 'shipjs trigger';
-    fs.writeFileSync(filePath, `${JSON.stringify(json, null, 2)}\n`);
-    runPrettier({ filePath, dir });
+export default async ({ dir, dryRun }) =>
+  await runStep({ title: 'Adding scripts to package.json' }, async () => {
+    if (!dryRun) {
+      const filePath = path.resolve(dir, 'package.json');
+      const json = JSON.parse(fs.readFileSync(filePath).toString());
+      json.scripts = json.scripts || {};
+      json.scripts['release:prepare'] = 'shipjs prepare';
+      json.scripts['release:trigger'] = 'shipjs trigger';
+      fs.writeFileSync(filePath, `${JSON.stringify(json, null, 2)}\n`);
+      await runPrettier({ filePath, dir });
+    }
     return () =>
       print(
         `${info(
